Revert "Revert "Merge branch 'kokkos:develop' into persistence""
This reverts commit 0bd01f5313f904c7e69984719c6893f7d3d0c714.
