+++ perf_tests/mpi/test_2dhalo.cpp
@@ -16,9 +16,9 @@
 
 #include "test_utils.hpp"
 
-#include "KokkosComm/KokkosComm.hpp"
-
 #include <iostream>
+
+#include <KokkosComm/KokkosComm.hpp>
 
 void noop(benchmark::State, MPI_Comm) {}
