+++ perf_tests/mpi/test_utils.hpp
@@ -18,9 +18,8 @@
 
 #include <chrono>
 
+#include <mpi.h>
 #include <benchmark/benchmark.h>
-
-#include "KokkosComm/mpi/impl/include_mpi.hpp"
 
 // F is a function that takes (state, MPI_Comm, args...)
 template <typename F, typename... Args>
