+++ src/KokkosComm/mpi/handle.hpp
@@ -16,8 +16,9 @@
 
 #pragma once
 
-#include "KokkosComm/fwd.hpp"
-
+#include <KokkosComm/fwd.hpp>
+#include <KokkosComm/concepts.hpp>
+#include "mpi_space.hpp"
 #include "req.hpp"
 
 namespace KokkosComm {
