+++ src/KokkosComm/mpi/impl/packer.hpp
@@ -16,10 +16,11 @@
 
 #pragma once
 
-#include "KokkosComm/concepts.hpp"
-#include "KokkosComm/traits.hpp"
+#include <mpi.h>
+
+#include <KokkosComm/concepts.hpp>
+#include <KokkosComm/traits.hpp>
 #include "types.hpp"
-#include "include_mpi.hpp"
 
 // todo: redo this using KokkosComm_contiguous
 
@@ -46,8 +47,6 @@ struct DeepCopy {
 
   template <KokkosExecutionSpace ExecSpace>
   static args_type allocate_packed_for(const ExecSpace &space, const std::string &label, const View &src) {
-    using KCT = KokkosComm::Traits<View>;
-
     if constexpr (KokkosComm::rank<View>() == 1) {
       non_const_packed_view_type packed(Kokkos::view_alloc(space, Kokkos::WithoutInitializing, label), src.extent(0));
       return args_type(packed, mpi_type_v<packed_value_type>, KokkosComm::span(packed));
@@ -83,8 +82,7 @@ struct MpiDatatype {
   template <KokkosExecutionSpace ExecSpace>
   static args_type allocate_packed_for(const ExecSpace & /*space*/, const std::string & /*label*/, const View &src) {
     using ValueType = typename View::value_type;
-
-    using KCT = KokkosComm::Traits<View>;
+    using KCT       = KokkosComm::Traits<View>;
 
     MPI_Datatype type = mpi_type<ValueType>();
     for (size_t d = 0; d < KokkosComm::Traits<View>::rank(); ++d) {
