+++ src/KokkosComm/mpi/send.hpp
@@ -16,18 +16,21 @@
 
 #pragma once
 
+#include <mpi.h>
 #include <Kokkos_Core.hpp>
 
-#include "commmode.hpp"
+#include <KokkosComm/concepts.hpp>
+#include <KokkosComm/traits.hpp>
+#include "comm_mode.hpp"
+
 #include "impl/pack_traits.hpp"
-#include "impl/include_mpi.hpp"
+#include "impl/types.hpp"
 
 namespace KokkosComm::mpi {
 
 template <KokkosView SendView, CommunicationMode SendMode>
 void send(const SendView &sv, int dest, int tag, MPI_Comm comm, SendMode) {
   Kokkos::Tools::pushRegion("KokkosComm::Impl::send");
-  using KCT = typename KokkosComm::Traits<SendView>;
 
   auto mpi_send_fn = [](void *mpi_view, int mpi_count, MPI_Datatype mpi_datatype, int mpi_dest, int mpi_tag,
                         MPI_Comm mpi_comm) {
@@ -49,6 +52,7 @@ void send(const SendView &sv, int dest, int tag, MPI_Comm comm, SendMode) {
   } else {
     throw std::runtime_error("only contiguous views supported for low-level send");
   }
+
   Kokkos::Tools::popRegion();
 }
