+++ unit_tests/mpi/test_gtest_mpi.cpp
@@ -16,7 +16,6 @@
 
 #include <gtest/gtest.h>
 #include <gtest/gtest-spi.h>
-
 #include <mpi.h>
 
 TEST(TestGtest, all_fail_nonfatal) {
