+++ unit_tests/test_sendrecv.cpp
@@ -15,8 +15,7 @@
 //@HEADER
 
 #include <gtest/gtest.h>
-
-#include "KokkosComm/KokkosComm.hpp"
+#include <KokkosComm/KokkosComm.hpp>
 
 #include "view_builder.hpp"
